assignments/movie-sentiment-aws/pyproject.toml
[project]
name = "movie-sentiment-aws"
version = "0.1.0"
description = "Serving a sentiment analysis model via FastAPI/Streamlit and deployed on AWS"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.14",
    "joblib>=1.5.1",
    "kagglehub>=0.3.12",
    "pandas>=2.3.0",
    "scikit-learn>=1.7.0",
    "uvicorn>=0.35.0",
    "streamlit>=1.46.0",
    "pyyaml>=6.0.1",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.12.0",
]

[tool.setuptools]
packages = ["src"]
